add authentication for Department Service and config swagger

+++ src/main/java/com/axonactive/jpa/controller/DepartmentController.java
@@ -1,8 +1,11 @@
 package com.axonactive.jpa.controller;
 
 import com.axonactive.jpa.controller.request.DepartmentRequest;
-import com.axonactive.jpa.service.DepartmentService;
+import com.axonactive.jpa.entities.Department;
+import com.axonactive.jpa.service.JWTAuthenticationServices;
 import com.axonactive.jpa.service.impl.DepartmentServiceImpl;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
@@ -11,41 +14,52 @@ import javax.ws.rs.core.Response;
 
 @Path("departments")
 @Produces(MediaType.APPLICATION_JSON)
+@Consumes({MediaType.APPLICATION_JSON,MediaType.APPLICATION_XML})
+@Api(value = "departments")
 public class DepartmentController {
 
     @Inject
     private DepartmentServiceImpl departmentService;
 
+    @Inject
+    JWTAuthenticationServices jwtAuthenticationServices;
+
     @GET
-    public Response getAllDepartments(){
+    @ApiOperation(value = "getAllDepartments", response =Department.class, responseContainer = "List")
+    public Response getAllDepartments(@HeaderParam("Authorization") String authorization){
+        jwtAuthenticationServices.checkAuthorizedToken(authorization);
         return Response.ok(departmentService.findAll()).build();
     }
 
 
     @GET
     @Path("/{id}")
-    public Response getDepartmentById(@PathParam("id") int id){
+    public Response getDepartmentById(@HeaderParam("Authorization") String authorization,@PathParam("id") int id){
+        jwtAuthenticationServices.checkAuthorizedToken(authorization);
         return Response.ok(departmentService.findById(id)).build();
     }
 
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response addDepartment(DepartmentRequest departmentRequest){
+    public Response addDepartment(@HeaderParam("Authorization") String authorization,DepartmentRequest departmentRequest){
+        jwtAuthenticationServices.checkAuthorizedToken(authorization);
         return Response.ok(departmentService.saveDepartment(departmentRequest)).build();
 
     }
 
     @DELETE
     @Path("/{id}")
-    public Response deleteDepartment(@PathParam("id") int id){
+    public Response deleteDepartment(@HeaderParam("Authorization") String authorization,@PathParam("id") int id){
+        jwtAuthenticationServices.checkAuthorizedToken(authorization);
         departmentService.deleteDepartment(id);
         return Response.ok().build();
     }
 
     @PUT
     @Path("/{id}")
-    public Response updateDepartment(@PathParam("id") int id, DepartmentRequest departmentRequest){
+    public Response updateDepartment(@HeaderParam("Authorization") String authorization,@PathParam("id") int id, DepartmentRequest departmentRequest){
+        jwtAuthenticationServices.checkAuthorizedToken(authorization);
         return Response.ok(departmentService.updateDepartment(id,departmentRequest)).build();
     }
